+++ apps/mobile/src/screens/(tabs)/cart.tsx
@@ -6,7 +6,9 @@ export default function CartScreen() {
   return (
     <View className='flex-1 bg-background'>
       <View className='container flex-1 items-center justify-center'>
-        <Text>Cart Screen</Text>
+        <Text variant='h5' className='text-muted-foreground'>
+          Your Cart is Empty
+        </Text>
       </View>
     </View>
   )